requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
docx==0.2.4
fpdf==1.7.2
idna==3.10
lxml==5.3.1
numpy==2.2.3
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
XlsxWriter==3.2.2
